[ refactor ] Relocating Goals & WarningsAndNonFatalErrors
