+++ src/full/Agda/Interaction/BasicOps.hs-boot
@@ -1,9 +1,5 @@
 module Agda.Interaction.BasicOps where
 
-import Agda.Syntax.Abstract (Expr)
-import Agda.Syntax.Common (InteractionId)
-import  {-# SOURCE #-} Agda.TypeChecking.Monad.Base (NamedMeta, TCWarning)
-
 data OutputForm a b
 data OutputConstraint a b
 data OutputConstraint' a b
@@ -19,11 +15,3 @@ instance Read ComputeMode
 data Rewrite
 instance Show Rewrite
 instance Read Rewrite
-
-type Goals = ( [OutputConstraint Expr InteractionId] -- visible metas
-             , [OutputConstraint Expr NamedMeta]     -- hidden metas
-             )
-type WarningsAndNonFatalErrors
-      = ( [TCWarning] -- warnings
-        , [TCWarning] -- non-fatal errors
-        )
