+++ src/full/Agda/Interaction/EmacsTop.hs-boot
@@ -1,0 +1,6 @@
+module Agda.Interaction.EmacsTop where
+
+import Agda.TypeChecking.Monad.Base (TCM)
+import Agda.Interaction.Response (Goals)
+
+showGoals :: Goals -> TCM String
