--- src/full/Agda/Interaction/InteractionTop.hs-boot
@@ -1,6 +1,0 @@
-module Agda.Interaction.InteractionTop where
-
-import Agda.TypeChecking.Monad.Base (TCM)
-import Agda.Interaction.BasicOps (Goals)
-
-showGoals :: Goals -> TCM String
